Add side-by-side git diff script

+++ install.sh
@@ -5,7 +5,7 @@ set -ex
 pwd=`pwd`
 
 for f in $(ls); do
-    if [[ $f != "install.sh" ]]; then
+    if [[ $f != "install.sh" && $f != "sbdiff.sh" ]]; then
         rm -rf $HOME/.$f
         ln -s $pwd/$f $HOME/.$f
     fi
